+++ packages/utils/src/constants/layer-zero-relayer-addresses.ts
@@ -1,6 +1,7 @@
 import { Address } from 'viem'
 import { Chains } from './chain-ids'
 
+// to find relayer address, check a tx on https://layerzeroscan.com/
 export const RELAYER_ADDRESSES: { [chain in Chains]?: Address } = {
   [Chains.ETHEREUM]: '0x902F09715B6303d4173037652FA7377e5b98089E',
   [Chains.OPTIMISM]: '0x81E792e5a9003CC1C8BF5569A00f34b65d75b017',
@@ -12,4 +13,5 @@ export const RELAYER_ADDRESSES: { [chain in Chains]?: Address } = {
   [Chains.ARBITRUM_ONE]: '0x177d36dBE2271A4DdB2Ad8304d82628eb921d790',
   [Chains.AVALANCHE]: '0xCD2E3622d483C7Dc855F72e5eafAdCD577ac78B4',
   [Chains.ZORA]: '0xA658742d33ebd2ce2F0bdFf73515Aa797Fd161D9',
+  [Chains.BLAST]: '0xA658742d33ebd2ce2F0bdFf73515Aa797Fd161D9',
 }